Separate the ShoppingCard componente into its own component
